requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
click==8.0.1
dlib==19.15.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.66.0
h11==0.12.0
jedi==0.18.0
multipart==0.2.4
numpy==1.21.0
parso==0.8.2
Pillow==8.3.1
pydantic==1.8.2
python-multipart==0.0.5
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
